+++ egui_demo_lib/src/apps/demo/plot_demo.rs
@@ -306,9 +306,9 @@ impl Widget for &mut LegendDemo {
 }
 
 #[derive(PartialEq, Default)]
-struct ItemsDemo {}
-
-impl ItemsDemo {}
+struct ItemsDemo {
+    texture: Option<egui::TextureHandle>,
+}
 
 impl Widget for &mut ItemsDemo {
     fn ui(self, ui: &mut Ui) -> Response {
@@ -343,12 +343,17 @@ impl Widget for &mut ItemsDemo {
             );
             Arrows::new(arrow_origins, arrow_tips)
         };
+
+        let texture: &egui::TextureHandle = self.texture.get_or_insert_with(|| {
+            ui.ctx()
+                .load_texture("plot_demo", egui::ColorImage::example())
+        });
         let image = PlotImage::new(
-            TextureId::Egui,
+            texture,
             Value::new(0.0, 10.0),
             [
-                ui.fonts().font_image().width as f32 / 100.0,
-                ui.fonts().font_image().height as f32 / 100.0,
+                ui.fonts().font_image().width() as f32 / 100.0,
+                ui.fonts().font_image().height() as f32 / 100.0,
             ],
         );