+++ egui_web/src/webgl2.rs
@@ -10,10 +10,7 @@ use {
     },
 };
 
-use egui::{
-    emath::vec2,
-    epaint::{Color32, FontImage},
-};
+use egui::{emath::vec2, epaint::Color32};
 
 type Gl = WebGl2RenderingContext;
 
@@ -28,13 +25,8 @@ pub struct WebGl2Painter {
     color_buffer: WebGlBuffer,
     post_process: PostProcess,
 
-    egui_texture: WebGlTexture,
-    egui_texture_version: Option<u64>,
-
-    /// Index is the same as in [`egui::TextureId::User`].
-    user_textures: HashMap<u64, WebGlTexture>,
-
-    next_native_tex_id: u64, // TODO: 128-bit texture space?
+    textures: HashMap<egui::TextureId, WebGlTexture>,
+    next_native_tex_id: u64,
 }
 
 impl WebGl2Painter {
@@ -85,18 +77,13 @@ impl WebGl2Painter {
             tc_buffer,
             color_buffer,
             post_process,
-            egui_texture,
-            egui_texture_version: None,
-            user_textures: Default::default(),
+            textures: Default::default(),
             next_native_tex_id: 1 << 32,
         })
     }
 
     fn get_texture(&self, texture_id: egui::TextureId) -> Option<&WebGlTexture> {
-        match texture_id {
-            egui::TextureId::Egui => Some(&self.egui_texture),
-            egui::TextureId::User(id) => self.user_textures.get(&id),
-        }
+        self.textures.get(&texture_id)
     }
 
     fn paint_mesh(&self, mesh: &egui::epaint::Mesh16) -> Result<(), JsValue> {
@@ -218,44 +205,8 @@ impl WebGl2Painter {
 
         Ok(())
     }
-}
 
-impl epi::NativeTexture for WebGl2Painter {
-    type Texture = WebGlTexture;
-
-    fn register_native_texture(&mut self, native: Self::Texture) -> egui::TextureId {
-        let id = self.next_native_tex_id;
-        self.next_native_tex_id += 1;
-        self.user_textures.insert(id, native);
-        egui::TextureId::User(id as u64)
-    }
-
-    fn replace_native_texture(&mut self, id: egui::TextureId, replacing: Self::Texture) {
-        if let egui::TextureId::User(id) = id {
-            if let Some(user_texture) = self.user_textures.get_mut(&id) {
-                *user_texture = replacing;
-            }
-        }
-    }
-}
-
-impl crate::Painter for WebGl2Painter {
-    fn set_texture(&mut self, tex_id: u64, image: epi::Image) {
-        assert_eq!(
-            image.size[0] * image.size[1],
-            image.pixels.len(),
-            "Mismatch between texture size and texel count"
-        );
-
-        // TODO: optimize
-        let mut pixels: Vec<u8> = Vec::with_capacity(image.pixels.len() * 4);
-        for srgba in image.pixels {
-            pixels.push(srgba.r());
-            pixels.push(srgba.g());
-            pixels.push(srgba.b());
-            pixels.push(srgba.a());
-        }
-
+    fn set_texture_rgba(&mut self, tex_id: egui::TextureId, size: [usize; 2], pixels: &[u8]) {
         let gl = &self.gl;
         let gl_texture = gl.create_texture().unwrap();
         gl.bind_texture(Gl::TEXTURE_2D, Some(&gl_texture));
@@ -271,24 +222,65 @@ impl crate::Painter for WebGl2Painter {
         let border = 0;
         let src_format = Gl::RGBA;
         let src_type = Gl::UNSIGNED_BYTE;
+        gl.pixel_storei(Gl::UNPACK_ALIGNMENT, 1);
         gl.tex_image_2d_with_i32_and_i32_and_i32_and_format_and_type_and_opt_u8_array(
             Gl::TEXTURE_2D,
             level,
-            internal_format as _,
-            image.size[0] as _,
-            image.size[1] as _,
+            internal_format as i32,
+            size[0] as i32,
+            size[1] as i32,
             border,
             src_format,
             src_type,
-            Some(&pixels),
+            Some(pixels),
         )
         .unwrap();
 
-        self.user_textures.insert(tex_id, gl_texture);
+        self.textures.insert(tex_id, gl_texture);
+    }
+}
+
+impl epi::NativeTexture for WebGl2Painter {
+    type Texture = WebGlTexture;
+
+    fn register_native_texture(&mut self, native: Self::Texture) -> egui::TextureId {
+        let id = egui::TextureId::User(self.next_native_tex_id);
+        self.next_native_tex_id += 1;
+        self.textures.insert(id, native);
+        id
     }
 
-    fn free_texture(&mut self, tex_id: u64) {
-        self.user_textures.remove(&tex_id);
+    fn replace_native_texture(&mut self, id: egui::TextureId, native: Self::Texture) {
+        self.textures.insert(id, native);
+    }
+}
+
+impl crate::Painter for WebGl2Painter {
+    fn set_texture(&mut self, tex_id: egui::TextureId, image: egui::ImageData) {
+        match image {
+            egui::ImageData::Color(image) => {
+                assert_eq!(
+                    image.width() * image.height(),
+                    image.pixels.len(),
+                    "Mismatch between texture size and texel count"
+                );
+
+                let data: &[u8] = bytemuck::cast_slice(image.pixels.as_ref());
+                self.set_texture_rgba(tex_id, image.size, data);
+            }
+            egui::ImageData::Alpha(image) => {
+                let gamma = 1.0;
+                let data: Vec<u8> = image
+                    .srgba_pixels(gamma)
+                    .flat_map(|a| a.to_array())
+                    .collect();
+                self.set_texture_rgba(tex_id, image.size, &data);
+            }
+        };
+    }
+
+    fn free_texture(&mut self, tex_id: egui::TextureId) {
+        self.textures.remove(&tex_id);
     }
 
     fn debug_info(&self) -> String {
@@ -305,44 +297,6 @@ impl crate::Painter for WebGl2Painter {
     /// id of the canvas html element containing the rendering
     fn canvas_id(&self) -> &str {
         &self.canvas_id
-    }
-
-    fn upload_egui_texture(&mut self, font_image: &FontImage) {
-        if self.egui_texture_version == Some(font_image.version) {
-            return; // No change
-        }
-
-        let mut pixels: Vec<u8> = Vec::with_capacity(font_image.pixels.len() * 4);
-        for srgba in font_image.srgba_pixels(1.0) {
-            pixels.push(srgba.r());
-            pixels.push(srgba.g());
-            pixels.push(srgba.b());
-            pixels.push(srgba.a());
-        }
-
-        let gl = &self.gl;
-        gl.bind_texture(Gl::TEXTURE_2D, Some(&self.egui_texture));
-
-        let level = 0;
-        let internal_format = Gl::SRGB8_ALPHA8;
-        let border = 0;
-        let src_format = Gl::RGBA;
-        let src_type = Gl::UNSIGNED_BYTE;
-        gl.pixel_storei(Gl::UNPACK_ALIGNMENT, 1);
-        gl.tex_image_2d_with_i32_and_i32_and_i32_and_format_and_type_and_opt_u8_array(
-            Gl::TEXTURE_2D,
-            level,
-            internal_format as i32,
-            font_image.width as i32,
-            font_image.height as i32,
-            border,
-            src_format,
-            src_type,
-            Some(&pixels),
-        )
-        .unwrap();
-
-        self.egui_texture_version = Some(font_image.version);
     }
 
     fn clear(&mut self, clear_color: egui::Rgba) {