+++ epaint/src/text/font.rs
@@ -376,12 +376,12 @@ fn allocate_glyph(
         } else {
             let glyph_pos = atlas.allocate((glyph_width, glyph_height));
 
-            let texture = atlas.image_mut();
+            let image = atlas.image_mut();
             glyph.draw(|x, y, v| {
                 if v > 0.0 {
                     let px = glyph_pos.0 + x as usize;
                     let py = glyph_pos.1 + y as usize;
-                    texture[(px, py)] = (v * 255.0).round() as u8;
+                    image.image[(px, py)] = (v * 255.0).round() as u8;
                 }
             });